requirements.txt
streamlit==1.37.0
python-docx==1.1.2
lxml==5.2.2
chromadb==0.5.5
langchain-community==0.2.10
langchain-text-splitters==0.2.2
sentence-transformers==3.0.1
pypdf==4.3.1
docx2txt==0.8
